Fix Bugs
- 실제 휴대폰 환경에서 화면 깨짐 현상 해결
- 마커 이동 후, 해당 좌표의 주소를 표시

+++ RegistGoodPlace/app/src/main/java/com/study/tedkim/registgoodplace/ShopInfo.java
@@ -10,7 +10,7 @@ import io.realm.annotations.PrimaryKey;
 public class ShopInfo extends RealmObject {
 
     @PrimaryKey
-    private int article_id;
+    private long article_id;
 
     private String name;
     private String address;
@@ -21,11 +21,11 @@ public class ShopInfo extends RealmObject {
 
     }
 
-    public int getArticle_id() {
+    public long getArticle_id() {
         return article_id;
     }
 
-    public void setArticle_id(int article_id) {
+    public void setArticle_id(long article_id) {
         this.article_id = article_id;
     }
